x86, jit: Inline caching optimization
This patch implements inline cache optimization that attempts to speed up
virtual method invocations with a 1-entry cache that is compiled into the code.
That is, instead of a vtable lookup, a call-site looks like this:

  movl #expected_klass, %eax
  call #method

We also add a small inline cache "landing zone" in the beginning of each
function which looks like this:

  cmpl	#this->class, %eax
  jne	#full_vtable_lookup

If we succesfully predict the target method, we simply continue executing it.
However, if this->class does not match, we take a full vtable lookup instead.

There's more information on the subject written by Cliff Click in his blog:

  http://www.azulsystems.com/blog/cliff-click/2010-04-08-inline-caches-and-call-site-optimization

This patch passes "make regression" but unfortunately I'm seeing errors with
DaCapo, for example. Help in tracking the issues down is much appreciated.

TODO:

  - Clean up inline_cache_supports_method

  - Instruction cache flushing

  - Interraction with vtable fixups

Signed-off-by: Pekka Enberg <penberg@kernel.org>