vm, verifier: Almost completely implementing the first verification pass.
The first phase of verification (as explicited in section 4.9.2, before
the data-flow analysis) is nearly completed. Are missing the analysis of
the offsets in the tableswitch and lookupswitch instructions, the
verification of the exception handlers limit and functionnal tests to
assert everything works fine (already hand-tested to some extent, and
passes make check).

Signed-off-by: Theo Dzierzbicki <theo.dz@gmail.com>

Added missing verifications.

Implemented the verifications of the offsets of tableswitch and
lookupswitch and checks of the exception table. Tests yet to be created.

Signed-off-by: Theo Dzierzbicki <theo.dz@gmail.com>

Added tests for the verifications implemented.

The tests are written now, though they still lack their rule in the
Makefile. Exception raising is not solved yet either.

Signed-off-by: Theo Dzierzbicki <theo.dz@gmail.com>