vm, verifier: Allow modification of blocks by verifier CFG functions.
When analyzing the basic blocks in the flow control analysis algorithm,
you don't know the state of the VM at the beginning of the block. As a
consequence, the "initial state" of the block contains an empty stack
and a local variables array with all vars initialized to UNKNOWN. But as
we go through the block we learn more about the state that we have to
expect at the beginning.

Previously the functions pop_vrf_op, push_vrf_op, ... were operating on
a verifier state (structure containing an array of local vars and a
stack) that corresponded to the "final state" of a block, so there was
no way to report for instance that a J_INT was expected in the initial
state stack of the block. Now they directly take verifier_blocks in
parameters, and modify both their initial state and final state.

Signed-off-by: Theo Dzierzbicki <theo.dz@gmail.com>